pyproject.toml
[project]
name = "digi-assist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.66.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "langgraph",
    "langchain",
]
authors = [
    {name = "Brijesh", email = "brij_joe@yahoo.com"},
]
license = {text = "GNU GENERAL PUBLIC LICENSE"}
